CHOP-32. Move from com.wibidata to org.kiji package.
