+++ pom.xml
@@ -3,7 +3,7 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.wibidata.chopsticks</groupId>
+  <groupId>org.kiji.chopsticks</groupId>
   <artifactId>kiji-chopsticks</artifactId>
   <version>0.1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
